add shell function ST_UD01 20140910

+++ ST_UD01/BSP/LED.h
@@ -12,7 +12,7 @@ extern CNT_TYPE  LED;
 
 /*** extern function prototype declarations ***/
 extern void LED_setup(void);
-extern void LED_Blink(uint8_t cnt);
+extern void LED_Blink(uint32_t cnt);
 
 #endif
